requirements.txt
streamlit
langchain
openai
langchain-openai 
python-dotenv